Revert "refactor[event]split logic and fix #149"
Revert "refactor[event]split logic and fix #149"

This reverts commit 1fe56f8a096afa979510df85e812660be8537f84.

Revert "fix[event]spill-the-tea check server-bind"

This reverts commit 73daafe6572e715d2beee3af3105fac98742d72c.

Revert "Merge pull request #269 from serfend/EventSubscript"

This reverts commit 2aeae9f1a3089c0b129ef93eaf80060e11e2dee6, reversing
changes made to 2ba52e323408b894e363fad64d76ac0670b5f37e.

Revert "fix[event]Serendipity check server-bind"

This reverts commit 78bbd0f04756f1e4bbb153b02785c9f9ebe5352c.

Revert "fix[event]typo register"

This reverts commit 7d829ed60175e67329a9f2418a996578156efb4e.

Revert "typo[event]wrong register"

This reverts commit 04a6e6c71f5a8922f366a6a1ce155350b12dfd28.
