+++ src/plugins/jx3/__init__.py
@@ -16,18 +16,25 @@ async def _():
 ws_recev = on(type="WsRecv", priority=5, block=False)
 @ws_recev.handle()
 async def on_ws_recev(bot: Bot, event: RecvEvent):
-    message = event.user_message_type
+    message = event.get_message()
     if not message:
         return
     groups = await bot.call_api("get_group_list")
     for i in groups:
         group = i["group_id"]
-        subscribe = json.loads(read(f"{DATA}{os.sep}{group}{os.sep}subscribe.json"))
-        if message in subscribe:
+        subscribe = json.loads(read(DATA + "/" + str(group) + "/subscribe.json"))
+        if message["type"] in subscribe:
             to_send_messagee = event.render_message(group)
-            if not to_send_messagee:
-                return # 无消息发送则退出
             try:
                 await bot.call_api("send_group_msg", group_id = group, message = to_send_messagee)
             except:
                 logger.info(f"向群({i})推送失败，可能是因为风控、禁言或者未加入该群。")
+
+# if message["type"] == "玄晶":
+#     group_info = json.loads(read(DATA + "/" + str(group) + "/jx3group.json"))
+#     if group_info["server"] != message["server"] and group_info["server"] != "":
+#         continue
+# elif message["type"] == "818":
+#     group_info = json.loads(read(DATA + "/" + str(group) + "/jx3group.json"))
+#     if group_info["server"] != "" and group_info["server"] != message["server"] and message["name"] != "剑网3":
+#         continue
