--- src/tools/dep/jx3/jx3apiws/jx3_event/__init__.py
@@ -1,2 +1,0 @@
-from .events import *
-from .BaseEvent import *
