--- src/tools/dep/jx3/jx3apiws/jx3_event/dep.py
@@ -1,1 +1,0 @@
-from .BaseEvent import *
