--- src/tools/dep/jx3/jx3apiws/jx3_event/events/HorseRefreshEvent.py
@@ -1,36 +1,0 @@
-from ..dep import *
-
-
-@EventRegister.register(action=1002)
-class HorseRefreshEvent(RecvEvent):
-    """马驹刷新事件"""
-
-    __event__ = "WsRecv.HorseRefresh"
-    message_type = "HorseRefresh"
-    user_message_type: str = '刷马'  # 用户定义的事件名称
-    map: str
-    """刷新地图"""
-    min: int
-    """时间范围min"""
-    max: int
-    """时间范围max"""
-    time: str
-    """推送时间"""
-
-    @validator("time", pre=True)
-    def check_time(cls, v):
-        start_trans = datetime.fromtimestamp(int(v))
-        return start_trans.strftime("%H:%M:%S")
-
-    @property
-    def log(self) -> str:
-        log = f"马驹刷新推送：[{self.server}]的[{self.map}]将要在 {str(self.min)}-{str(self.max)} 分后刷新马驹。"
-        return log
-
-    @overrides(RecvEvent)
-    def render_message(self, group) -> str:
-        return self.get_message()
-
-    @overrides(RecvEvent)
-    def get_message(self) -> str:
-        f"[抓马监控] 时间：{self.time}\n{self.map} 将在[{self.min} - {self.max}分]后刷新马驹。"
