--- src/tools/dep/jx3/jx3apiws/jx3_event/events/__init__.py
@@ -1,14 +1,0 @@
-from .ClientUpdateEvent import *
-from .DisSubscribeEvent import *
-from .FireworksEvent import *
-from .FuyaoNamedEvent import *
-from .FuyaoRefreshEvent import *
-from .GameSysMsgEvent import *
-from .HorseCatchedEvent import *
-from .HorseRefreshEvent import *
-from .NewsEvent import *
-from .SerendipityEvent import *
-from .ServerStatusEvent import *
-from .SpillTheTeaEvent import *
-from .SubscribeEvent import *
-from .XuanJingEvent import *
